[fixup] make CloudCollectorStartStopCaptureRequestWaiter externally available
