+++ src/CaptureServiceBase/include/CaptureServiceBase/CloudCollectorStartStopCaptureRequestWaiter.h
@@ -5,13 +5,30 @@
 #ifndef CAPTURE_SERVICE_BASE_CLOUD_COLLECTOR_START_STOP_CAPTURE_REQUEST_WAITER_H_
 #define CAPTURE_SERVICE_BASE_CLOUD_COLLECTOR_START_STOP_CAPTURE_REQUEST_WAITER_H_
 
+#include <absl/base/thread_annotations.h>
+#include <absl/synchronization/mutex.h>
+
 #include "CaptureServiceBase/StartStopCaptureRequestWaiter.h"
+#include "GrpcProtos/capture.pb.h"
 
 namespace orbit_capture_service_base {
 
-// Create a `StartStopCaptureRequestWaiter` for the cloud collector.
-[[nodiscard]] std::shared_ptr<StartStopCaptureRequestWaiter>
-CreateCloudCollectorStartStopCaptureRequestWaiter();
+// A `StartStopCaptureRequestWaiter` implementation for the cloud collector.
+class CloudCollectorStartStopCaptureRequestWaiter : public StartStopCaptureRequestWaiter {
+ public:
+  [[nodiscard]] orbit_grpc_protos::CaptureOptions WaitForStartCaptureRequest() override;
+  void StartCapture(orbit_grpc_protos::CaptureOptions capture_options);
+
+  void WaitForStopCaptureRequest() override;
+  void StopCapture();
+
+ private:
+  mutable absl::Mutex capture_options_mutex_;
+  CaptureOptions capture_options_ ABSL_GUARDED_BY(capture_options_mutex_);
+
+  std::atomic<bool> start_requested_ = false;
+  std::atomic<bool> stop_requested_ = false;
+};
 
 }  // namespace orbit_capture_service_base
